+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_inputs_outputs/output.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-# pylint: disable=redefined-builtin
+# pylint: disable=redefined-builtin, too-many-instance-attributes
 import re
 from typing import Dict, overload